// src/bot/models/bot.model.ts
import { Column, DataType, Model, Table } from "sequelize-typescript";

interface IBotCreationAtrr {
  user_id: number;
  username: string;
  first_name: string;
  last_name: string;
  lang: string;
}
@Table({ tableName: "bot" })
export class Bot extends Model<Bot, IBotCreationAtrr> {
  @Column({ type: DataType.BIGINT, primaryKey: true })
  user_id: number;

  @Column({ type: DataType.STRING })
  username: string;

  @Column({ type: DataType.STRING })
  first_name: string;

  @Column({ type: DataType.STRING })
  last_name: string;

  @Column({ type: DataType.STRING })
  phone_number: string;

  @Column({ type: DataType.STRING })
  lang: string;

  @Column({ type: DataType.BOOLEAN, defaultValue: false })
  status: boolean;
}